PHP/Trait Use Statements: minor tweak to "correct" example
Allowing a single conflict resolution statement on the same line as the `use` decreases readability.

Changing this to a multi-line statement, with each conflict resolution statement on its own line, also improves consistency with other WP Core rules around the use of braces, as it makes the brace layout consistent with the expected brace layout for various control structures.