Update html.md
The XHTML specification was superseded in 2018, and the HTML specification is now considered the standard of reference. 

As such, while we can still reference the XHTML specification, there are places where XHTML standards conflict with HTML standards, such as the handling of boolean attributes. We should update the HTML standards to explicitly allow boolean attributes. See https://core.trac.wordpress.org/ticket/60178 for reference.