Update accessibility.md
Update the Accessibility Coding Standards and related references to WCAG 2.2. WCAG 2.2 was established as the most recent version of the official W3C recommendations in October 2023, and it is past time that we migrate our standards to coordinate. They are already established in law in the UK, and the European Accessibility Act that comes into enforcement in June 2025 has revisions in progress to update those requirements.